+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### 1.0.0-alpha2 - 2025-07-01
+
 ### Added
 
 - Add support for multiple logs in herbicide quick form.